Formatting and cleaning (#91)
* automatic cleaning and rename methods and packages to match convention

* add IntelliJ auto formatting files
