--- src/main/java/com/BrewMES/demo/model/Report.java
@@ -1,11 +1,0 @@
-package com.BrewMES.demo.model;
-
-import java.io.File;
-
-public class Report {
-
-	public static File generatePDF(Batch batch) {
-		throw new UnsupportedOperationException();
-	}
-
-}
