+++ src/main/java/com/brewmes/demo/BrewMesApplication.java
@@ -1,6 +1,5 @@
-package com.BrewMES.demo;
+package com.brewmes.demo;
 
-import com.BrewMES.demo.model.BrewMES;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
