+++ src/main/java/com/brewmes/demo/api/StringResponse.java
@@ -1,4 +1,5 @@
-package com.BrewMES.demo.API;
+package com.brewmes.demo.api;
+
 //Class for wrapping string responses in a simple object assosiating the http response code.
 public class StringResponse {
 
@@ -14,12 +15,12 @@ public class StringResponse {
         return respose;
     }
 
-    public int getStatus() {
-        return status;
-    }
-
     public void setRespose(String respose) {
         this.respose = respose;
+    }
+
+    public int getStatus() {
+        return status;
     }
 
     public void setStatus(int status) {
