+++ src/main/java/com/brewmes/demo/model/Command.java
@@ -1,4 +1,4 @@
-package com.BrewMES.demo.model;
+package com.brewmes.demo.model;
 
 public enum Command {
     RESET(1),
