+++ src/main/java/com/brewmes/demo/model/MachineState.java
@@ -1,4 +1,4 @@
-package com.BrewMES.demo.model;
+package com.brewmes.demo.model;
 
 public enum MachineState {
     DEACTIVATED(0),
