+++ src/main/java/com/brewmes/demo/model/Report.java
@@ -1,0 +1,11 @@
+package com.brewmes.demo.model;
+
+import java.io.File;
+
+public class Report {
+
+    public static File generatePDF(Batch batch) {
+        throw new UnsupportedOperationException();
+    }
+
+}
