+++ src/test/java/com/brewmes/demo/BrewMesApplicationTests.java
@@ -1,4 +1,4 @@
-package com.BrewMES.demo;
+package com.brewmes.demo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
